+++ src/modules/lists/module.js
@@ -186,7 +186,7 @@ const listsModule = {
     },
   },
   category: "core-workflow",
-  version: "0.33.5.20.5",
+  version: "0.33.5.20.6",
   enabledByDefault: true,
   canDisable: true,
   historicalReadAccess: true,